+++ libretro-db/rmsgpack_dom.c
@@ -474,7 +474,7 @@ int rmsgpack_dom_read_into(RFILE *fd, ...)
             min_len     = (value->val.binary.len > *uint_value) ?
                *uint_value : value->val.binary.len;
 
-            memcpy(buff_value, value->val.binary.buff, min_len);
+            memcpy(buff_value, value->val.binary.buff, (size_t)min_len);
             break;
          case RDT_STRING:
             buff_value = va_arg(ap, char *);
@@ -483,7 +483,7 @@ int rmsgpack_dom_read_into(RFILE *fd, ...)
                *uint_value : value->val.string.len + 1;
             *uint_value = min_len;
 
-            memcpy(buff_value, value->val.string.buff, min_len);
+            memcpy(buff_value, value->val.string.buff, (size_t)min_len);
             break;
          default:
             rv = -1;
